% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cssl.R
\name{cssl}
\alias{cssl}
\title{Quantitative trait loci (QTLs) mapping for quantitative trait in
chromosome segment substitution lines (CSSLs)}
\usage{
cssl(
  genotype = genotype,
  phenotype = phenotype,
  dire = NULL,
  cores = 1,
  LOD = 2.5,
  BLUP = FALSE,
  qq = FALSE
)
}
\arguments{
\item{genotype}{The parameter allows three file input formats.
\itemize{
\item Character string with path of the genotype file;
\item Data frame of the genotype;
\item Matrix of the genotype.
}

See details in vignette.}

\item{phenotype}{The parameter allows three file input formats.
\itemize{
\item Character string with path of the phenotype file;
\item Data frame of the phenotype;
}

See details in vignette.}

\item{dire}{Character string with path for saving results. Defaults value is
\code{NULL} and will save in working directory, using \link{getwd} to get the
working directory.}

\item{cores}{The number of CPU cores to use, for parallel calculations.
Defaults is 1 and no parallel.}

\item{LOD}{A numerical value indicates the threshold of LOD. Defaults is
\code{2.5}.}

\item{BLUP}{Defaults is \code{FALSE}. If it is \code{TRUE}, will calculate
the BLUP (best linear unbiased prediction) values for trait measured in
multi-environment.}

\item{qq}{Defaults is \code{FALSE}. If it is \code{TRUE}, will transform
quantitative trait phenotypic value using function \link[stats:qqnorm]{stats::qqnorm}. It is
useful for expression quantitative trait locus (eQTL) mapping.}
}
\value{
A list include two data frame.

\item{map}{The map have four columns, including:}
\itemize{
\item Chromosome, the names of chromosome
\item Start, the position of start
\item End, the position of end
\item Marker, the names of marker
}

\item{result}{The result have eight columns, including:}
\itemize{
\item Trait, the trait names
\item Environment, the environment names
\item Marker, the marker names
\item Effect, the values of effect
\item Type, "additive" or "dominant"
\item LOD, the LOD value
\item Same_marker, the code of Same_marker is same as the Marker
\item r^2(\%), the ratio of QTL variance and total variance
}
}
\description{
The multi-locus linear model for quantitative traits in CSSLs
is described as follow
\deqn{y_i = \mu + \sum_{j=1}^m{\omega_{ij}a_j} + \sum_{j=1}^m{\nu_{ij}d_j} +
 \varepsilon_i} where the \eqn{y_i} is the phenotypic observation values
of quantitative trait of the \emph{i}th line (\eqn{i = 1, 2, \cdots, n}); \eqn{\mu}
is total average; \eqn{a_j} and \eqn{d_j} are the additive and dominance
effects of the \emph{j}th locus, \eqn{\omega_{ij}} and \eqn{\nu_{ij}} are
their corresponding indicator variables. We assumed that the allele A and a
of each locus are derived from donor and background parents, so
\deqn{\omega_{ij}=\begin{cases}1 & AA \\ 0 & Aa \\ -1 & aa \end{cases}}
and \deqn{\nu_{ij}=\begin{cases}0 & AA \\ 1 & Aa \\ 0 & aa \end{cases}}
}
\examples{
# library(cssl)
\dontrun{
#' data(SimData)
result <- cssl(genotype  = SimData$genotype,
               phenotype = SimData$phenotype,
               LOD       = 5)
}
}
