% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimData}
\alias{SimData}
\title{An example of CSSLs}
\format{
The list of \code{SimData} has 3 elements:
\describe{
\code{genotype} is a data frame.

\code{phenotype} is a data frame.

\code{geno} is a matrix.
}
}
\usage{
SimData
}
\description{
\code{SimData} is a data set of CSSLs. It is a list include
\code{genotype}, \code{phenotype} and \code{geno}.
}
\examples{
# library(mppQTL)
data(SimData)
}
\keyword{datasets}
